ech: filter all HTTPS records
